Add correct resetting of screen for next number
A global variable, newNum, holds a boolean value that tells
numButtonPress when to reset the screen value. This value is set to true
when an operator button is pressed, so the previous number stays on the
screen until the next number is entered. Once a new number has been
started, numButtonPress resets newNum to a false value.