Add divide by zero handling and C button function
Function equals now checks for 'ERR' as a possible value and passes it
to the screen without running other checks. div returns this string if
zero is passed as its second operand.

C and CE buttons now have distinct functions. C clears the screen and
the current operation, CE just clears the screen but leaves any existing
operation in place.