Get input display working
Probably not done yet but a lot is done. Added a new field to
InputBuffer, it now keeps track of a seperate list of symbols that
are display-friendly. Modified Calculator's logic for determining what
to use for keyValue and keySymbol to be more readable and use the new
data field properly.

Added a list of inputs that are functions and have Calculator
automatically add an open parenthesis after.

Changed control property names from secondaryLabel
/ secondaryValue to secondLabel / secondaryValue.